cart cards adjustment
