+++ src/components/Cart/index.jsx
@@ -1,70 +1,8 @@
-import React, { useContext, useState } from 'react';
-import { toast } from "react-toastify";
+import React, { useContext, useState } from "react";
 import CartContext from "../../context/CartContext";
 import { roundWithTwoDecimals } from "../../helpers/functions";
+import CartCard from "./CartCard";
 import "./style.scss";
-
-const CartCard = ({ cartItem, editCart }) => {
-  const { updateQuantity, removeFromCart } = useContext(CartContext);
-  const { title, quantity, image, id } = cartItem;
-
-  const increaseQuantity = () => {
-    const newQuantity = quantity + 1;
-    updateQuantity({ id, quantity: newQuantity });
-  };
-
-  const decreaseQuantity = () => {
-    const newQuantity = quantity - 1;
-    if (newQuantity === 1) {
-      toast.warning("Removing the remaing item, will remove it from cart");
-    }
-    updateQuantity({ id, quantity: newQuantity });
-  };
-
-  const handleChange = ({ target }) => {
-    if (target.value === 1) {
-      toast.warning("Removing the remaing item, will remove it from cart");
-    }
-    updateQuantity({ id, quantity: target.value });
-  };
-
-  const handleRemove = () => {
-    removeFromCart({ id });
-  };
-
-  return (
-    <div className="cart-card">
-      <div className="cart-card-image">
-        <img src={image} alt={title} />
-      </div>
-      <div className="cart-card-content">
-        <h5>
-          {quantity}x {title}
-        </h5>
-
-        {editCart && (
-          <div className="edit-cart">
-            <button onClick={handleRemove} className="btn-absolute">
-              &#10005;
-            </button>
-            <button onClick={decreaseQuantity} className="btn">
-              -
-            </button>
-            <input
-              value={quantity}
-              type="number"
-              min="0"
-              onChange={handleChange}
-            />
-            <button onClick={increaseQuantity} className="btn">
-              +
-            </button>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
 
 const Cart = ({ closeCart }) => {
   const { cartTotal, shoppedItems } = useContext(CartContext);
@@ -77,7 +15,7 @@ const Cart = ({ closeCart }) => {
 
   return (
     <div className="cart-container">
-      <span onClick={closeCart}>Close Cart</span>
+      <span onClick={closeCart}>&#10005;</span>
       <div className="cart-header">
         <h3>Order Summary</h3>
         <p>{shoppedItems.length} items</p>
@@ -140,4 +78,4 @@ const Cart = ({ closeCart }) => {
   );
 };
 
-export default Cart
+export default Cart;
